RF: Move FontManager to `tools` so it can be imported without importing `visual`
