+++ psychopy/visual/textbox2/textbox2.py
@@ -29,10 +29,8 @@ from ..basevisual import (
 )
 from psychopy.tools.attributetools import attributeSetter, setAttribute
 from psychopy.tools import mathtools as mt
-from psychopy.tools.arraytools import val2array
-from psychopy.tools.monitorunittools import convertToPix
 from psychopy.colors import Color
-from .fontmanager import FontManager, GLFont
+from psychopy.tools.fontmanager import FontManager, GLFont
 from .. import shaders
 from ..rect import Rect
 from ... import core, alerts, layout
